+++ src/gauge/SpeedTape.h
@@ -7,19 +7,19 @@
   This project is distributed under the terms of the GNU General Public License
   Version 3 <http://www.gnu.org/licenses/gpl.html>.
   
-      This program is free software: you can redistribute it and/or modify
-      it under the terms of the GNU General Public License as published by
-      the Free Software Foundation, specifically version 3 of the License.
+  This program is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, specifically version 3 of the License.
   
-      This program is distributed in the hope that it will be useful,
-      but WITHOUT ANY WARRANTY; without even the implied warranty of
-      MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-      GNU General Public License for more details.
+  This program is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
   
-      You should have received a copy of the GNU General Public License
-      along with this program.  If not, see <http://www.gnu.org/licenses/>.
+  You should have received a copy of the GNU General Public License
+  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-=========================================================================*/
+  =========================================================================*/
 
 /**
  * Boeing 777 style speed tape
@@ -33,23 +33,23 @@
 namespace OpenGC
 {
 
-class SpeedTape : public GaugeComponent  
-{
-	public:
+  class SpeedTape : public GaugeComponent  
+  {
+  public:
 
-		SpeedTape();
-		virtual ~SpeedTape();
+    SpeedTape();
+    virtual ~SpeedTape();
 
-		/** Overloaded render function */
-		void Render();
+    /** Overloaded render function */
+    void Render();
 
-	protected:
+  protected:
 
-		/** The font number provided to us by the font manager */
-		int m_Font;
+    /** The font number provided to us by the font manager */
+    int m_Font;
 
-		double indent_x;  // defines region of speed tape
-};
+    float indent_x;  // defines region of speed tape
+  };
 
 } // end namespace OpenGC
