+++ src/gauge/Tachometer.cpp
@@ -54,20 +54,22 @@ namespace OpenGC
   void Tachometer::RenderTicks(CircleEvaluator *circ)
   {
     // yellow stripe
-    double percentagey = m_MinYellow / (m_Max - m_Min) ;
-    double degreeyellow = minDegrees + ((maxDegrees - minDegrees) * percentagey);
+    float percentagey = m_MinYellow / (m_Max - m_Min) ;
+    float degreeyellow = minDegrees + ((maxDegrees - minDegrees) * percentagey);
     glColor3ub(247, 231, 8);
-    double radians = degreeyellow * DEG_TO_RAD;
+    float radians = degreeyellow * DEG_TO_RAD;
+    
     const float vertices[] = {
       R * sin(radians), R * cos(radians),
       (R + 4) * sin(radians), (R + 4) * cos(radians)
     };
+    
     glVertexPointer(2, GL_FLOAT, 0, &vertices);
     glDrawArrays(GL_LINE_STRIP, 0, 2);
 
     // red stripe
-    double percentager = m_MinRed / (m_Max - m_Min) ;
-    double degreered =  minDegrees + ((maxDegrees - minDegrees) * percentager);
+    float percentager = m_MinRed / (m_Max - m_Min) ;
+    float degreered =  minDegrees + ((maxDegrees - minDegrees) * percentager);
     radians = degreered * DEG_TO_RAD;
     glColor3ub(255, 0, 0);
     const float vertices2[] = {
